requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
gTTS==2.4.0
pydub==0.25.1
reportlab==4.0.7
python-dotenv==1.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0